requirements.txt
# Required for local models
llama-cpp-python>=0.2.0

# Optional for API models
openai>=1.0.0
anthropic>=0.7.0